+++ src/rpls.rs
@@ -18,7 +18,7 @@ use crate::barcode;
 pub fn pairs_to_landscape(bd_pairs: Vec<BirthDeath>, k:usize, debug:bool) -> Result<Vec<Vec<persistencelandscape::PointOrd>>, &'static str>{
     let bd_pairs: Vec<BirthDeath> = bd_pairs
         .into_iter()
-        .filter(|bd| (bd.birth - bd.death).abs() > f64::EPSILON)
+        .filter(|bd| (bd.birth - bd.death).abs() > f32::EPSILON)
         .collect();
     if bd_pairs.is_empty() {
         if debug {
@@ -41,7 +41,7 @@ pub fn pairs_to_landscape(bd_pairs: Vec<BirthDeath>, k:usize, debug:bool) -> Res
     Ok(landscape)
 }
 
-fn area_under_line_segment(a: persistencelandscape::PointOrd, b: persistencelandscape::PointOrd) ->f64 {
+fn area_under_line_segment(a: persistencelandscape::PointOrd, b: persistencelandscape::PointOrd) ->f32 {
     let height = (a.y.0 - b.y.0).abs();
     let base = a.x.0 - b.x.0;
     let triangle = (height * base) / 2.0;
@@ -55,12 +55,12 @@ fn area_under_line_segment(a: persistencelandscape::PointOrd, b: persistenceland
     triangle + rectangle
 }
 
-fn landscape_norm(landscape: &[persistencelandscape::PointOrd]) -> f64 {
+fn landscape_norm(landscape: &[persistencelandscape::PointOrd]) -> f32 {
     landscape
         .iter()
         .zip(landscape.iter().skip(1))
         .map(|(&a, &b)| area_under_line_segment(a, b))
-        .sum::<f64>()
+        .sum::<f32>()
 }
 
 fn is_sorted<T>(data: &[T]) -> bool
@@ -74,11 +74,11 @@ where
 ///
 /// Will panic if areas are not strictly decreasing or equal
 #[must_use]
-pub fn l2_norm(landscapes: &[Vec<persistencelandscape::PointOrd>]) -> f64 {
+pub fn l2_norm(landscapes: &[Vec<persistencelandscape::PointOrd>]) -> f32 {
     let areas = landscapes
         .iter()
         .map(|l| FloatOrd(landscape_norm(l)))
-        .collect::<Vec<FloatOrd<f64>>>();
+        .collect::<Vec<FloatOrd<f32>>>();
     assert!(is_sorted(& areas));
 
         areas.iter().map(|x| x.0).sum()
@@ -87,6 +87,6 @@ pub fn l2_norm(landscapes: &[Vec<persistencelandscape::PointOrd>]) -> f64 {
 /// # Errors
 ///
 /// Will return 'Err' if failed to compute persistencelandscape from `bd_pairs`
-pub fn pairs_to_l2_norm(bd_paris: Vec<BirthDeath>, k:usize, debug:bool) -> Result<f64, &'static str>{
+pub fn pairs_to_l2_norm(bd_paris: Vec<BirthDeath>, k:usize, debug:bool) -> Result<f32, &'static str>{
     Ok(l2_norm(pairs_to_landscape(bd_paris, k, debug)?.as_slice()))
 }
